Pass through XX-Cf-app-instance as x-cf-app-instance
We've seen weird behaviour (that's difficult to reproduce) where
binding a route service to a route causes the x-cf-app-instance header
to fail to route to the requested instance.  We suspect that this is
because:

 - when you have a route service, your request goes through gorouter
   twice
 - on the first pass through, the x-cf-app-instance header is consumed
   and stripped, then your request goes to the route service
 - on the second pass through, the header is missing, so gorouter
   randomly picks an instance to send the request to

As an ugly workaround, this commit allows you to specify
XX-cf-app-instance instead, which will not be stripped on the first
pass through gorouter, then the route service sets x-cf-app-instance
from this header, so that the second pass through gorouter will get
the required value in the x-cf-app-instance header and route to the
correct instance.