Update route service to work with nginx buildpack
The staticfile_buildpack no longer supports custom nginx configuration,
as such this repo is not in a deployable state.

The makefiles are a bit confusing and we can rely on the PaaS
documentation instead.

We can rely on the manifest being a template, and using the `--var`
features of CF push as well.

I've also updated the nginx configuration to include the correct http
version when proxying.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>