Avoid newlines in nginx config
Using \n characters as separators for allow statements in the nginx config file
causes a syntax error and prevents the buildpack from being built. But removing the
newlines works and results in the allow lines to be all on the same line, which
doesn't substantially reduce the readability of the file.